added slugs back in for 404 and home pages

+++ App/models/LeadboxesModel.php
@@ -21,49 +21,54 @@ class LeadboxesModel
         Security::checkAdminRefererStatic('save_leadbox_options');
 
         $globalLeadboxes = array(
-            'lp_select_field_0' => sanitize_text_field($_POST['lp_select_field_0']),
-            'leadboxes_timed_display_radio' => sanitize_text_field((!empty($_POST['leadboxes_timed_display_radio']) ? $_POST['leadboxes_timed_display_radio'] : '' )),
-            'lp_select_field_2' => sanitize_text_field($_POST['lp_select_field_2']),
-            'leadboxes_exit_display_radio' => sanitize_text_field((!empty($_POST['leadboxes_exit_display_radio']) ? $_POST['leadboxes_exit_display_radio'] : '' ))
+          'lp_select_field_0'             => sanitize_text_field($_POST['lp_select_field_0']),
+          'leadboxes_timed_display_radio' => sanitize_text_field((!empty($_POST['leadboxes_timed_display_radio']) ? $_POST['leadboxes_timed_display_radio'] : '')),
+          'lp_select_field_2'             => sanitize_text_field($_POST['lp_select_field_2']),
+          'leadboxes_exit_display_radio'  => sanitize_text_field((!empty($_POST['leadboxes_exit_display_radio']) ? $_POST['leadboxes_exit_display_radio'] : ''))
         );
 
         static::updateLeadboxOption($globalLeadboxes);
-        wp_redirect(admin_url().'?page=Leadboxes');
+        wp_redirect(admin_url() . '?page=Leadboxes');
     }
 
-    protected static function updateLeadboxOption($data){
+    protected static function updateLeadboxOption($data)
+    {
         update_option('lp_settings', $data);
     }
 
-    public static function getLpSettings(){
+    public static function getLpSettings()
+    {
         $data = get_option('lp_settings');
-        if(!is_wp_error($data)){
+        if (!is_wp_error($data)) {
             return $data;
         }
     }
 
-    public static function getCurrentTimedLeadbox(){
+    public static function getCurrentTimedLeadbox()
+    {
         $leadboxes = static::getLpSettings();
-        if(!empty($leadboxes['lp_select_field_0'])){
+        if (!empty($leadboxes['lp_select_field_0'])) {
             $currentTimedLeadbox = array($leadboxes['lp_select_field_0'], $leadboxes['leadboxes_timed_display_radio']);
-        }else{
+        } else {
             $currentTimedLeadbox = array('none', 'none');
         }
         return $currentTimedLeadbox;
     }
 
-    public static function getCurrentExitLeadbox(){
+    public static function getCurrentExitLeadbox()
+    {
         $leadboxes = static::getLpSettings();
-        if(!empty($leadboxes['lp_select_field_2'])){
+        if (!empty($leadboxes['lp_select_field_2'])) {
             $currentExitLeadbox = array($leadboxes['lp_select_field_2'], $leadboxes['leadboxes_exit_display_radio']);
-        }else{
+        } else {
             $currentExitLeadbox = array('none', 'none');
         }
         return $currentExitLeadbox;
     }
 
 
-    public static function savePageSpecificLeadboxes($post_id, $post){
+    public static function savePageSpecificLeadboxes($post_id, $post)
+    {
 
         self::savePageSpecificTimedLeadbox($post_id);
 
@@ -82,13 +87,15 @@ class LeadboxesModel
      */
     public static function savePageSpecificTimedLeadbox($post_id)
     {
-        $timedLeadbox = sanitize_text_field($_POST['pageTimedLeadbox']);
+        if (isset($_POST['pageTimedLeadbox'])) {
+            $timedLeadbox = sanitize_text_field($_POST['pageTimedLeadbox']);
 
-        if ($timedLeadbox != 'select') {
-            update_post_meta($post_id, 'pageTimedLeadbox', $timedLeadbox);
-        } else {
-            //if switched back to select delete the post meta so global leadboxs will display again
-            delete_post_meta($post_id, 'pageTimedLeadbox');
+            if ($timedLeadbox != 'select') {
+                update_post_meta($post_id, 'pageTimedLeadbox', $timedLeadbox);
+            } else {
+                //if switched back to select delete the post meta so global leadboxs will display again
+                delete_post_meta($post_id, 'pageTimedLeadbox');
+            }
         }
     }
 
@@ -98,13 +105,15 @@ class LeadboxesModel
      */
     public static function savePageSpecificExitLeadbox($post_id)
     {
-        $exitLeadbox  = sanitize_text_field($_POST['pageExitLeadbox']);
+        if (isset($_POST['pageExitLeadbox'])) {
+            $exitLeadbox = sanitize_text_field($_POST['pageExitLeadbox']);
 
-        if ($exitLeadbox != 'select') {
-            update_post_meta($post_id, 'pageExitLeadbox', $exitLeadbox);
-        } else {
-            //if switched back to select delete the post meta so global leadboxs will display again
-            delete_post_meta($post_id, 'pageExitLeadbox');
+            if ($exitLeadbox != 'select') {
+                update_post_meta($post_id, 'pageExitLeadbox', $exitLeadbox);
+            } else {
+                //if switched back to select delete the post meta so global leadboxs will display again
+                delete_post_meta($post_id, 'pageExitLeadbox');
+            }
         }
     }
